requirements.txt
flask==2.1.0
tensorflow==2.10.0
opencv-python==4.6.0.66
numpy==1.23.4
werkzeug==2.1.1
pillow==9.2.0 